app/serializers/product.py
from rest_framework import serializers

from app.models import Product

class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = ('id','product_name','product_code','category')