requirements.txt
asgiref==3.5.2
autopep8==2.0.0
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.4
django-cors-headers==3.13.0
django-filter==22.1
django-mongodb-engine==0.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.0
djangotoolbox==1.8.0
djongo==1.3.6
dnspython==2.2.1
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==22.0
pycodestyle==2.10.0
PyJWT==2.6.0
pymongo==3.12.3
python-snappy==0.6.1
pytz==2022.6
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.2.4
tomli==2.0.1
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
winkerberos==0.9.1
